web_app/requirements.txt
Flask==3.1.0
tensorflow==2.20.0
numpy>=1.26.0
Pillow>=10.1.0
gunicorn>=21.2.0
Werkzeug>=3.0.1